refactor: 기능별 역활 패키지 명칭 분리
